cloudfunction/requirements.txt
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
oauth2client==4.1.3
requests==2.23.0
